Refs #2897: int cast in for loops

+++ Code/Mantid/Framework/Algorithms/src/AbsorptionCorrection.cpp
@@ -125,7 +125,7 @@ void AbsorptionCorrection::exec()
   Progress prog(this,0.0,1.0,numHists);
   // Loop over the spectra
   PARALLEL_FOR2(m_inputWS,correctionFactors)
-  for (int64_t i = 0; i < numHists; ++i)
+  for (int64_t i = 0; i < int64_t(numHists); ++i)
   {
     PARALLEL_START_INTERUPT_REGION
