+++ Code/Mantid/Framework/Algorithms/src/SetUncertaintiesToZero.cpp
@@ -63,7 +63,7 @@ void SetUncertaintiesToZero::exec()
   Progress prog(this,0.0,1.0,numHists);
 
   PARALLEL_FOR2(inputWorkspace,outputWorkspace)
-  for (int i = 0; i < numHists; ++i)
+  for (int64_t i = 0; i < int64_t(numHists); ++i)
   {
     PARALLEL_START_INTERUPT_REGION
